Add classes for CompoundEntry and ReactionEntry instead of namedtuples
This makes the classes more similar to the other CompoundEntry and
ReactionEntry classes that are implemented for datasources in the main
model_script which makes it easier to merge these implementations into
one later. It also makes it possible to load models from model_script
datasources and pass them through model-import for conversion to the
native format.