model: Add method _try_parse_formula()
With this change, the importers will try to parse the compound formulas
and will provide an error in case the formula could not be parsed.
However, the original formula string is copied directly to the YAML model
files in order to preserve the original formula.

This also fixes importing the formula for a number of Excel models where
the neutral formula was used instead of the charged formula.