importer: Merge equivalent compounds
Add option to merge compound entries that represent the same
compound but in various different compartments. Some models
are specified in terms of compound pools where the compartment
is an inherent part of the entity. In this case, there are
multiple entries for the same compound in various compartments.
With this option, we try to merge these so that the resulting
model better complies with the convention for PSAMM models,
that compounds in diverse compartments should be represented
by a single compound entry.