model: Change MetabolicModel to a mutable object
This change makes it much easier to apply various modifications
to a model during import (e.g. conversion of exchange reactions
to exchange compound definition etc.) This was not an issue
initially but as the functionality of psamm-import has expanded
it has become necessary to import models in multiple phases
and applying updates or modifications in multiple passes. This
is also the first step towards creating a shared model
representation that can be used by both PSAMM and psamm-import.