importer: Do not complain about dest when list is requested
When the main command is run with "list" as argument, the command
should just list the possible model formats. However, if the
current directory is not empty, the command instead fails with
an error that the directory must be empty. This error message
should only be shown after the user has selected a model.